Adjusted the file format and key generation alg.
Now the file has a version number, and stores the
original filename so that it doesn't matter
what you call the encrypted file.  For keygen, it
now starts with the IV, to better thwart attempts
to take shortcuts when brute-forcing the key.
It also hashes the working state an additional 5000
times so that it takes more computation to try a
password.