Autotools is in effect again, and hashing is working.
I'm on my way to recombining the functions into a single `spritz`
binary, like the Go implementation.  I'm also taking the time
to redo the functions to allocate less--callers are expected
to supply the buffers, which helps to re-use memory across
invocations.

Next will come the scaffolding to select a sub-command, and then
the encryption functions will be brought up to date.