Added old-style encryption, and encryption test script.
The old-style encryption didn't do extensive keygen, didn't
store the "real" encryption key so that a file can be re-keyed,
and didn't store the original file name.  The next step is to
enhance the encryption to do these things.  It will probably NOT
be fully compatible with the Go version, but very similar.