Added the embedded filenames and non-decrypting check.
Now the filename is embedded in the file, so that you don't have
to store the encryped file with a descriptive name anymore.

Also, if you pass '-n' instead of '-d' to the crypt command, you
will just get a readout of what the output filename would be, which
incidentally tells you if you got the password right.