A big simplification: no more outfile handling.
Instead of doing a "just okay" job at handling output directories
and embedding filenames, I'm leaving that to `tar`.  That way,
you also get owners and permissions, etc.  It's just better for
everyone.  The only downside is that you can't encrypt more than
one file in the same invocation of the program, but that's
a rare and minor issue (usually encrypting a lot of files starts
with tarring them up into a single file anyway.)